Moved unexported methods to manager.go.
